Testedinsim

+++ src/barometer.hpp
@@ -28,11 +28,11 @@ class Barometer {
 
     public:
 
-        void    init(const BarometerConfig & _config, Board * _board);
-        void    calibrate(void);
-        void    update(void);
-        int32_t getAltitude(void);
-        int32_t getVelocity(uint32_t dTimeMicros);
+        void  init(const BarometerConfig & _config, Board * _board);
+        void  calibrate(void);
+        void  update(void);
+        float getAltitude(void);
+        float getVelocity(uint32_t dTimeMicros);
 
     private:
 
@@ -40,15 +40,15 @@ class Barometer {
 
         Board * board;
 
-        int32_t  alt;
-        int32_t  history[BarometerConfig::HISTORY_SIZE];
-        int      historyIdx;
-        int32_t  groundPressure;
-        int32_t  groundAltitude;
-        int32_t  lastAlt;
-        uint32_t pressureSum;
+        float   alt;
+        float   history[BarometerConfig::HISTORY_SIZE];
+        uint8_t historyIdx;
+        float   groundPressure;
+        float   groundAltitude;
+        float   lastAlt;
+        float   pressureSum;
 
-        static int32_t paToCm(uint32_t pa);
+        static float paToCm(float pa);
 };
 
 /********************************************* CPP ********************************************************/
@@ -66,7 +66,7 @@ void Barometer::init(const BarometerConfig & _config, Board * _board)
     alt = 0;
     lastAlt = 0;
 
-    for (int k=0; k<BarometerConfig::HISTORY_SIZE; ++k) {
+    for (uint8_t k=0; k<BarometerConfig::HISTORY_SIZE; ++k) {
         history[k] = 0;
     }
 }
@@ -83,25 +83,25 @@ void Barometer::update()
     float pressure = board->extrasGetBaroPressure();
 
     uint8_t indexplus1 = (historyIdx + 1) % BarometerConfig::HISTORY_SIZE;
-    history[historyIdx] = (int32_t)pressure;
+    history[historyIdx] = pressure;
     pressureSum += history[historyIdx];
     pressureSum -= history[indexplus1];
     historyIdx = indexplus1;
 }
 
-int32_t Barometer::getAltitude(void)
+float Barometer::getAltitude(void)
 {
-    int32_t alt_tmp = paToCm(pressureSum/(BarometerConfig::HISTORY_SIZE-1)); 
+    float alt_tmp = paToCm(pressureSum/(BarometerConfig::HISTORY_SIZE-1)); 
     alt_tmp -= groundAltitude;
-    alt = lrintf((float)alt * config.noiseLpf + (float)alt_tmp * (1.0f - config.noiseLpf));
+    alt = lrintf(alt * config.noiseLpf + alt_tmp * (1.0f - config.noiseLpf));
 
     return alt;
 }
 
-int32_t Barometer::getVelocity(uint32_t dTimeMicros)
+float Barometer::getVelocity(uint32_t dTimeMicros)
 {
-    static int32_t lastAlt;
-    int32_t vel = (int32_t)((alt - lastAlt) * 1000000.0f / dTimeMicros);
+    static float lastAlt;
+    float vel = (alt - lastAlt) * 1000000.0f / dTimeMicros;
     lastAlt = alt;
     vel = Filter::constrainAbs(vel, config.velocityBound); 
     return Filter::deadband(vel, config.velocityDeadband);
@@ -109,9 +109,9 @@ int32_t Barometer::getVelocity(uint32_t dTimeMicros)
 
 
 // Pressure in Pascals to altitude in centimeters
-int32_t Barometer::paToCm(uint32_t pa)
+float Barometer::paToCm(float pa)
 {
-    return (int32_t)((1.0f - powf(pa / 101325.0f, 0.190295f)) * 4433000.0f);
+    return (1.0f - powf(pa / 101325.0f, 0.190295f)) * 4433000.0f;
 }
 
 } // namespace hf
